item lar  tuzatildi

+++ src/main/java/sarik/dev/foodwaveproject/entity/OrderItem.java
@@ -25,7 +25,7 @@ public class OrderItem {
 	private Order order;
 	
 	private Integer quantity;
-	private double discount;
-	private double orderedProductPrice;
+	private Long discount;
+	private Long orderedProductPrice;
 	
 }
